Exception handling

+++ src/main/java/com/SpringProfessional/dscommerce/services/ProductService.java
@@ -5,14 +5,20 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.SpringProfessional.dscommerce.dto.ProductDTO;
 import com.SpringProfessional.dscommerce.entities.Product;
 import com.SpringProfessional.dscommerce.repositories.ProductRepository;
+import com.SpringProfessional.dscommerce.services.exceptions.DataBaseException;
+import com.SpringProfessional.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -22,37 +28,53 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado!"));
 		return new ProductDTO(product);
 	}
-	
+
 	@Transactional(readOnly = true)
-	public Page<ProductDTO> findAll(Pageable pageable){
-		Page<Product> result=repository.findAll(pageable);
-	    return result.map(x-> new ProductDTO(x));
-	    
+	public Page<ProductDTO> findAll(Pageable pageable) {
+		Page<Product> result = repository.findAll(pageable);
+		return result.map(x -> new ProductDTO(x));
+
 	}
-	
+
 	@Transactional
 	public ProductDTO insert(ProductDTO dto) {
-		Product entity=new Product();
+		Product entity = new Product();
 		CopyDtoToEntity(dto, entity);
-		entity=repository.save(entity);
+		entity = repository.save(entity);
 		return new ProductDTO(entity);
-		
+
 	}
-	
+
 	@Transactional
-	public ProductDTO update(Long id,ProductDTO dto) {
-		Product entity=repository.getReferenceById(id);
-		CopyDtoToEntity(dto,entity);
-		entity=repository.save(entity);
-		return new ProductDTO(entity);
+	public ProductDTO update(Long id, ProductDTO dto) {
+		try {
+			Product entity = repository.getReferenceById(id);
+			CopyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		} 
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
+
 	}
-	
-	@Transactional
+
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de Integridade Referencial!");
+		}
+	
 	}
 
 	private void CopyDtoToEntity(ProductDTO dto, Product entity) {
@@ -61,6 +83,5 @@ public class ProductService {
 		entity.setPrice(dto.getPrice());
 		entity.setImgUrl(dto.getImgUrl());
 	}
-
 
 }
